requirements.txt
bcrypt==4.3.0
bleach==6.2.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Deprecated==1.2.18
Flask==3.1.1
Flask-Bcrypt==1.0.1
Flask-Limiter==3.12
Flask-Mail==0.10.0
Flask-WTF==1.2.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
Pygments==2.19.2
pymysql==1.1.1
python-dotenv==1.1.1
rich==13.9.4
typing_extensions==4.14.1
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.3
WTForms==3.2.1
supabase==1.0.0